requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9 # Нужно для приема файлов (UploadFile) и форм

sqlalchemy>=2.0.28
alembic>=1.13.1
asyncpg>=0.29.0
greenlet==3.3.0

celery>=5.3.6
redis>=5.0.3

groq>=0.4.2

aioboto3==15.5.0
gspread>=6.0.0
gspread_asyncio==2.0.0
google-auth>=2.28.0

aiofiles==25.1.0
httpx>=0.27.0
python-dotenv>=1.0.1